onGazeLong repeats after complete
When using Fuse and onGazeLong after the fuse time is complete the fuse is i still visible and keeps firing the onGazeLong callback every cycle of time set on the fuse time. This is causing that all actions triggered on this callback keeps refireing and do not a allow to simulate click with th fuse one you execute an action only once.